Don't build separate image for migration.
The migration container uses the same image as the app, but before
this change a new image was being built off the same docker
container. This moves migraitons to both depend on app and use it's
image.
It also mounts the source into the volume in case the dev is testing
changes to migrations or the ingest script.