reorganize code into submodules (#106)
* move api code to submodule

* add symlink

* fix symlink

* fastapi_stac -> stac_fastapi

* update symlink

* add app routes, update imports

* rename

* add backend submodule

* add symlink

* update imports

* add extensions submodule, move core extensions from api

* add extensions symlink

* move bulk transactions to third party extensions

* move tiles extension to third_party

* backend.client -> backend.core

* define third party clients next to the extension

* remove third party clients from core

* correct package names

* create postgres submodule, move the sqlalchemy code

* move config

* update extension imports

* move errors

* move openapi

* fix tests

* add types submodule for shared types, finish moving db code, save progress on seperation of concerns...

* lint everything, remove mypy for now

* update pytest.ini

* Updates to submodules PR (#109)

* Consolidate image building to only use docker-compose

The makefile was building a named image, which was the same image as
the docker-compose app service image, but they were not the same
name. This commit consolidates the building and running of the image
through the app service of the docker-compose, and absorbes the
.env.example file into the env setup in docker-compose.yml

* Use docker network instead of external hosts.

* Add stac_fastapi_server package.

This brings in the logic of stac_api.app into it's own
subpackage. This is used by the development environment to spin up
stac-fastapi via docker-compose

* Read joplin data locally.

The remotely hosted joplin data was failing the ingest as they do not
have a "links" property, which makes them invalid STAC - the previous
server didn't complain, but that was breaking the ingest into
stac-fastapi.

This adds an empty link properties to the data downloaded from the remote location.

* Add fields to ApiSettings.

This pulls in the settings fields taken from
stac_api.config.ApiSettings.

* Don't build separate image for migration.

The migration container uses the same image as the app, but before
this change a new image was being built off the same docker
container. This moves migraitons to both depend on app and use it's
image.
It also mounts the source into the volume in case the dev is testing
changes to migrations or the ingest script.

* Pip install subpackages in the Dockerfile

* Structure dependencies in setup.py's

Restructure the Dockerfile to copy only what it needs as it installs
the subpackages. This will help catch dependency errors.

* Delete old stac_api package

* Add version to stac-fastapi-server

* Add a publish script.

This will be used by CI to publish on tags.

* Add publish GitHub action.

This will publish stac-fastapi to pypi on any tag.

* Fix broken import

* Pull top level package version from api version

* Pin stac-pydantic to 1.3.8

* Move search to polygon method out of types.

Has dependency on shapely, which can be avoided in the types subproject.

* Use Optional in param as it's possible to pass None

* Refactor set logic

* Pin sqlalchemy to 1.3.23

* Move global settings registry into stac-fastapi.types

Settings are used in the types subproject, which is the base
layer. This change makes the global instance of the settings that is
set by the API to be held in the types subproject to avoid circular dependencies.

* Move logic to include query fields in Fields to postgres client.

This logic requires settings that are specific to the postgres
client. Move this logic into the postgres client subpackage and remove
from the pydantic validator chain.

* Fixes from pre-commit formatting and linting

* lint

* rename postgres package to sqlalchemy

* rename PostgresSettings to SqlalchemySettings

* update readme

* relock types pipfile

Co-authored-by: Rob Emanuele <rdemanuele@gmail.com>