Use total=False for TypedDict
TypedDicts require that either all attributes are considered required,
or None. Because STAC has optional attributes that we don't want users
to put "None" into (and therefore make the serialized responses have
"null" values in fields, which produces invalid STAC), it's better to
allow all fields to be optional. This allows users to utilize the type
information in the TypedDict when constructing STAC objects, but
not be told by type hinting to supply optional fields they don't
have. This produces some type errors around direct dict access to
fields of the potential for the dict key to not be present.