feat: remove links from default include list
**sqlalchemy** uses `FieldsExtension.default_includes`, so its fix was pretty
*easy
-- I did have to add an explicit include+exclude check to the search function to
ensure the fields extension was applied _only if_ either `include` or `exclude`
was set. **pgstac** was a little trickier, because it doesn't use the
`default_includes` list.

I did modify one existing test -- `test_app_fields_extension` in the
**sqlalchemy** tests was assuming fields would be stripped just by providing a
`collections` parameter to the query, which I think was incorrect -- I added an
explicit `include` field to the test.